changed how display type is performed
Now each plottable has a display type parameter that controls how to display the line
A regression will therefore never be a marker, and if dashed is specified it can make itself dashed
if errors are specified, always are markers. A regression is always a line type. Specifying dashed and xy errors without a regression is an error.